# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NairaMaster.ng Micro Task Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/me/delete:
    post:
      operationId: deleteMyAccount
      tags: [auth]
      summary: Delete own account permanently
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountInput"
      responses:
        "200":
          description: Account deleted

  # --- WALLET ---
...
  /submissions/{submissionId}/screenshots:
    post:
      operationId: uploadScreenshots
      tags: [submissions]
      summary: Upload screenshots for a submission (max 2)
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenshotUploadInput"
      responses:
        "200":
          description: Screenshots uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{submissionId}/send:
    post:
      operationId: sendSubmission
      tags: [submissions]
      summary: Send submission to admin for review
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # --- WITHDRAWALS ---
...
  /docs:
    get:
      operationId: getDocs
      tags: [documentation]
      summary: Get all documentation
      responses:
        "200":
          description: Documentation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doc"
    post:
      operationId: createDoc
      tags: [documentation]
      summary: Create documentation (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doc"
  /docs/{docId}:
    patch:
      operationId: updateDoc
      tags: [documentation]
      summary: Update documentation (admin)
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doc"
    delete:
      operationId: deleteDoc
      tags: [documentation]
      summary: Delete documentation (admin)
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # --- UPDATES / ANNOUNCEMENTS ---
...
  /banners:
    get:
      operationId: getBanners
      tags: [banners]
      summary: Get active banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [banners]
      summary: Create a banner (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /banners/{bannerId}:
    delete:
      operationId: deleteBanner
      tags: [banners]
      summary: Delete a banner (admin)
      parameters:
        - name: bannerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # --- NOTIFICATIONS ---
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read

  # --- SETTINGS ---
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [firstName, lastName, email, gender, password]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        gender:
          type: string
          enum: [male, female, other]
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    DeleteAccountInput:
      type: object
      required: [confirmed]
      properties:
        confirmed:
          type: boolean
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, firstName, lastName, email, gender, balance, score, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        gender:
          type: string
        balance:
          type: number
        score:
          type: integer
        status:
          type: string
          enum: [active, blocked, suspended, shadow_banned, restricted, disabled]
        statusReason:
          type: ["string", "null"]
        statusUntil:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        restrictions:
          type: ["object", "null"]
        dailyDuesPaid:
          type: boolean
        createdAt:
          type: string
        lastLoginAt:
          type: ["string", "null"]

    # --- Wallet ---
...
    Task:
      type: object
      required: [id, title, description, link, socialMedia, category, tags, maxPerformers, performerCount, status, uploadedByAdmin, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        link:
          type: string
        socialMedia:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        maxPerformers:
          type: integer
        performerCount:
          type: integer
        status:
          type: string
          enum: [active, expired, deleted]
        uploadedByAdmin:
          type: boolean
        uploadedByUserId:
          type: ["integer", "null"]
        reward:
          type: ["number", "null"]
        performedByCurrentUser:
          type: boolean
        createdAt:
          type: string
...
    Submission:
      type: object
      required: [id, taskId, userId, status, createdAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, declined]
        screenshots:
          type: array
          items:
            type: string
        declineReason:
          type: ["string", "null"]
        task:
          $ref: "#/components/schemas/Task"
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ScreenshotUploadInput:
      type: object
      required: [screenshots]
      properties:
        screenshots:
          type: array
          items:
            type: string
          maxItems: 2

    # --- Withdrawals ---
...
    Doc:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        linkUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    DocInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        linkUrl:
          type: string

    # --- Updates ---
...
    Banner:
      type: object
      required: [id, title, content, active, intervalMinutes, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        ctaText:
          type: ["string", "null"]
        ctaLink:
          type: ["string", "null"]
        intervalMinutes:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
    BannerInput:
      type: object
      required: [title, content, intervalMinutes]
      properties:
        title:
          type: string
        content:
          type: string
        ctaText:
          type: string
        ctaLink:
          type: string
        intervalMinutes:
          type: integer

    # --- Notifications ---
...